Update definitions for react-onclickoutside 6
