+++ types/react-onclickoutside/index.d.ts
@@ -1,38 +1,35 @@
-// Type definitions for react-onclickoutside 5.7
+// Type definitions for react-onclickoutside 6.0
 // Project: https://github.com/Pomax/react-onclickoutside
 // Definitions by: Karol Janyst <https://github.com/LKay>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 2.1
 
-import * as React from 'react';
+import * as React from "react";
 
-declare namespace OnClickOut {
-    interface HandleClickOutside<T> {
-        handleClickOutside: React.MouseEventHandler<T>;
-    }
+export interface HandleClickOutside<T> {
+    handleClickOutside: React.MouseEventHandler<T>;
+}
 
-    interface InjectedOnClickOutProps {
-        disableOnClickOutside(): void;
-        enableOnClickOutside(): void;
-    }
+export interface InjectedOnClickOutProps {
+    disableOnClickOutside(): void;
+    enableOnClickOutside(): void;
+}
 
-    interface OnClickOutProps {
-        disableOnClickOutside?: boolean;
-        eventTypes?: string | string[];
-        outsideClickIgnoreClass?: string;
-        preventDefault?: boolean;
-        stopPropagation?: boolean;
-    }
+export interface OnClickOutProps {
+    disableOnClickOutside?: boolean;
+    eventTypes?: string | string[];
+    outsideClickIgnoreClass?: string;
+    preventDefault?: boolean;
+    stopPropagation?: boolean;
 }
 
 type ComponentConstructor<P> = React.ComponentClass<P> | React.StatelessComponent<P>;
-interface ClickOutComponentClass<P extends OnClickOut.InjectedOnClickOutProps> extends React.ComponentClass<P> {
-    new (props?: P, context?: any): React.Component<P, React.ComponentState> & OnClickOut.HandleClickOutside<any>;
+
+interface ClickOutComponentClass<P extends InjectedOnClickOutProps> extends React.ComponentClass<P> {
+    new (props?: P, context?: any): React.Component<P, React.ComponentState> & HandleClickOutside<any>;
 }
 
-declare function OnClickOut<P>(
-    component: ComponentConstructor<P & OnClickOut.InjectedOnClickOutProps & OnClickOut.HandleClickOutside<any>>
-        | ClickOutComponentClass<P & OnClickOut.InjectedOnClickOutProps>
-): React.ComponentClass<P & OnClickOut.OnClickOutProps>;
-
-export = OnClickOut;
+export default function OnClickOut<P>(
+    component: ComponentConstructor<P & InjectedOnClickOutProps & HandleClickOutside<any>>
+        | ClickOutComponentClass<P & InjectedOnClickOutProps>
+): React.ComponentClass<P & OnClickOutProps>;
