+++ types/react-onclickoutside/v5/tsconfig.json
@@ -1,0 +1,6 @@
+{
+    "extends": "../tsconfig.json",
+    "files": [
+        "index.d.ts"
+    ]
+}
